Refactor date formatter functions to utils module
Add auto-generated API docs for date formatter functions so that they are
discoverable for model profile creators.
